requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
spacy
torch
transformers
faiss-cpu==1.7.4
phonenumbers
